+++ src/main/java/roomescape/reservation/QueryingDAO.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
