+++ src/main/java/roomescape/reservation/UpdatingDAO.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
