+++ src/main/java/roomescape/reservationtime/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservationtime;
 
 import java.time.LocalTime;
